% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtree.R
\name{crosstabToCases}
\alias{crosstabToCases}
\title{crosstabToCases}
\usage{
crosstabToCases(x)
}
\arguments{
\item{x}{a matrix or table of frequencies representing a crosstabulation.}
}
\value{
Returns a data frame of cases.
}
\description{
Convert a crosstabulation into a data frame of cases.
}
\author{
Nick Barrowman
}
