% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FakeRCT}
\alias{FakeRCT}
\title{Fake Randomized Controlled Trial (RCT) data}
\format{
A small data frame in which the rows represent (imaginary) patients and
the columns represent variables of possible clinical relevance.

\describe{
  \item{id}{String: Patient ID number}
  \item{eligible}{Factor: Eligible or Ineligible}
  \item{randomized}{Factor: Randomized or Not randomized}
  \item{group}{Factor: A or B}
  \item{followup}{Factor: Followed up or Not followed up}
  \item{analyzed}{Factor: Analyzed or Not analyzed}

}
}
\usage{
FakeRCT
}
\description{
A dataset consisting of made-up RCT data.
}
\keyword{datasets}
