% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{init_js}
\alias{init_js}
\title{Initializing JS-Part}
\usage{
init_js(init_event, onwindow_resize, shortcuts)
}
\arguments{
\item{init_event}{The mouse event to activate zooming and paning. Default is \code{mouseenter}.}

\item{onwindow_resize}{Should the SVG be resized when the window size changes?
Default is \code{TRUE}.}

\item{shortcuts}{Should Keyboard shortcuts be used to control the SVG?
Default is \code{TRUE}.}
}
\description{
Initializing JS-Part
}
\seealso{
Other Shiny Functions: \code{\link{inlineCssSetup}},
  \code{\link{renderVtree}}, \code{\link{use_svgzoom}},
  \code{\link{vtreeOutput}}
}
\concept{Shiny Functions}
\keyword{internal}
