% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{inlineCssSetup}
\alias{inlineCssSetup}
\title{Setup for inline CSS}
\usage{
inlineCssSetup(minheight, cursor_all, overflow, position, fill,
  cursor_text)
}
\arguments{
\item{minheight}{minimum height in \code{"px"}. Default is "200px".}

\item{cursor_all}{The cursor symbol for the whole SVG. Default is "all-scroll".}

\item{overflow}{Overflow value for the whole SVG. Default is "inherit".}

\item{position}{CSS position of the SVG. Default is "sticky".}

\item{fill}{Fill color for the SVG background. Default is "transparent".}

\item{cursor_text}{The cursor symbol for text nodes. Default is "pointer".}
}
\description{
Initial CSS setup for the SVG.
}
\seealso{
\code{\link{use_svgzoom}}

Other Shiny Functions: \code{\link{init_js}},
  \code{\link{renderVtree}}, \code{\link{use_svgzoom}},
  \code{\link{vtreeOutput}}
}
\concept{Shiny Functions}
\keyword{internal}
