% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtree.R
\name{nodeids}
\alias{nodeids}
\title{Node function: IDs}
\usage{
nodeids(u, varname, value, args)
}
\arguments{
\item{u}{Data frame}

\item{varname}{The name of the variable in this node}

\item{value}{The value of the variable in this node.}

\item{args}{A list of arguments}
}
\description{
Node function: IDs
}
\author{
Nick J. Barrowman
}
