% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtree.R
\name{nodeMeanSD}
\alias{nodeMeanSD}
\title{Node function: mean and SD}
\usage{
nodeMeanSD(u, varname, value, args)
}
\arguments{
\item{u}{Data frame}

\item{varname}{The name of the variable in this node}

\item{value}{The value of the variable in this node.}

\item{args}{A list of arguments}
}
\description{
Node function: mean and SD
}
\author{
Nick J. Barrowman
}
