% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtree.R
\name{ptableExtra}
\alias{ptableExtra}
\title{ptableExtra}
\usage{
ptableExtra(x, markdown = FALSE)
}
\arguments{
\item{x}{Pattern table from vtree for indicator (i.e 0/1) variables}

\item{markdown}{Format nicely for markdown (see details).}
}
\value{
Returns a character matrix with extra rows containing indicator sums.
}
\description{
\code{ptableExtra} Augument a pattern table of indicator (i.e 0/1) variables that 
                    was produced by vtree (by specifying \code{ptable=TRUE})
}
\details{
The column totals ignore missing values.

When \code{markdown=TRUE}, the row and column headings for percentages are 
labeled "%", indicator values equal to 1 are replaced by checkmark codes,
indicator values equal to 0 are replaced by spaces, and missing indicator
values are replaced by dashes. Empty headings are replaced by spaces.
Finally the table is trasposed.
}
\examples{

ptableExtra(tab)

}
\author{
Nick Barrowman
}
