% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vtree.R
\docType{package}
\name{vtree-package}
\alias{vtree-package}
\title{vtree: a tool for calculating and drawing variable trees.}
\description{
vtree is a flexible tool for generating variable trees —
diagrams that display information about nested subsets of a data frame.
Given simple specifications,
the \code{vtree} function produces these diagrams and automatically
labels them with counts, percentages, and other summaries.

With vtree, you can:
\itemize{
  \item explore a data set interactively, and 
  \item produce customized figures for reports and publications.
}

For a comprehensive introduction, type: \code{vignette("vtree")}
}
\seealso{
\itemize{
  \item \url{https://nbarrowman.github.io/vtree}
  \item \url{https://github.com/nbarrowman/vtree}
  \item Report bugs at \url{https://github.com/nbarrowman/vtree/issues}
}
}
\author{
Nick Barrowman <nbarrowman@cheo.on.ca>
}
