% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svtree.R
\name{svtree}
\alias{svtree}
\title{Create a shiny vtree, with svg-pan-zoom functionality.}
\usage{
svtree(...)
}
\arguments{
\item{...}{parameters to be passed to `vtree`}
}
\description{
Create a shiny vtree, with svg-pan-zoom functionality.
}
